Reorganized file structure and removed extraneous files.

+++ src/lifegrid.cpp
@@ -24,11 +24,11 @@ LifeGrid::LifeGrid(int width, int height, int side, std::vector<sf::Color> color
     if(threads == 0)
         calcSums(0,cells.size());
     else{
-        for(int i = 0; i < threads; i++){
+        for(int i = 0; i < (int) threads; i++){
             activeThreads.push_back(std::thread(&LifeGrid::calcSums, this, i/threads*cells.size(), (i+1)/threads*cells.size()));
         }
     }
-    for(int i = 0; i < activeThreads.size(); i++){
+    for(int i = 0; i < (int) activeThreads.size(); i++){
         activeThreads[i].join();
     }
     activeThreads.resize(0);
@@ -41,11 +41,11 @@ void LifeGrid::update(){
     if(threads == 0)
         calcSums(0,cells.size());
     else{
-        for(int i = 0; i < threads; i++){
+        for(int i = 0; i < (int) threads; i++){
             activeThreads.push_back(std::thread(&LifeGrid::updateCells, this, i/threads*cells.size(), (i+1)/threads*cells.size()));
         }
     }
-    for(int i = 0; i < activeThreads.size(); i++){
+    for(int i = 0; i < (int) activeThreads.size(); i++){
         activeThreads[i].join();
     }
 
@@ -54,11 +54,11 @@ void LifeGrid::update(){
     if(threads == 0)
         calcSums(0,cells.size());
     else{
-        for(int i = 0; i < threads; i++){
+        for(int i = 0; i < (int) threads; i++){
             activeThreads.push_back(std::thread(&LifeGrid::calcSums, this, i/threads*cells.size(), (i+1)/threads*cells.size()));
         }
     }
-    for(int i = 0; i < activeThreads.size(); i++){
+    for(int i = 0; i < (int) activeThreads.size(); i++){
         activeThreads[i].join();
     }
 }
